+++ threadpoolctl.py
@@ -11,13 +11,16 @@ maximal number of threads they can use.
 # adapted from code by Intel developer @anton-malakhov available at
 # https://github.com/IntelPython/smp (Copyright (c) 2017, Intel Corporation)
 # and also published under the BSD 3-Clause license
+
+from __future__ import annotations
+
 import os
 import re
 import sys
 import ctypes
 import itertools
 import textwrap
-from threading import Thread
+from threading import Lock, Thread
 from typing import Callable, Literal, final
 import warnings
 from abc import ABC, abstractmethod
@@ -570,6 +573,11 @@ class FlexiBLASController(LibController):
                     " a backend available in the FlexiBLAS configuration "
                     f"{self.available_backends} or the path to a valid shared library."
                 )
+
+            # Invalidate the global cached controller:
+            global _CACHED_CONTROLLER
+            with _CACHE_LOCK:
+                _CACHED_CONTROLLER = None
 
             # Trigger a new search of loaded shared libraries since loading a new
             # backend caused a dlopen.
@@ -1737,6 +1745,37 @@ def _main():
         exec(options.command)
 
     print(json.dumps(threadpool_info(debugging_info=True), indent=2))
+
+
+# A cached instance of ThreadpoolController, created by
+# get_cached_controller():
+_CACHED_CONTROLLER = None
+# Cached number of modules in sys.modules, used by get_cached_controller()
+_CACHED_SYS_MODULES_LEN = 0
+# The lock controlling access to the above:
+_CACHE_LOCK = Lock()
+
+
+def get_cached_controller() -> ThreadpoolController:
+    """Return a cached ``ThreadpoolController``, as a speed optimization.
+
+    The cached version is regenerated if the number of imported of modules has
+    changed, since this indicates the possibility that a new controllable
+    library has been loaded.  Invalidation also happens if Flexiblas is in use
+    and has changed its backend.
+
+    It is possible that somehow a controllable library gets loaded without
+    importing a Python module, e.g. using ``dlopen()`` on Linux, a ctypes
+    ``CDLL``, and the like.  This is rare, however.  Outside of these
+    situations, using ``get_cached_controller()`` will be much faster than
+    creating a new ``ThreadpoolController`` instance.
+    """
+    global _CACHED_CONTROLLER, _CACHED_SYS_MODULES_LEN
+    with _CACHE_LOCK:
+        if _CACHED_CONTROLLER is None or len(sys.modules) != _CACHED_SYS_MODULES_LEN:
+            _CACHED_CONTROLLER = ThreadpoolController()
+            _CACHED_SYS_MODULES_LEN = len(sys.modules)
+    return _CACHED_CONTROLLER
 
 
 if __name__ == "__main__":